requirements/core.txt
psycopg2==2.5.1
-e git+https://github.com/django/django.git@1.6c1#egg=django
djangorestframework==2.3.8
python3-memcached==1.51
South==0.8.2
